Move an QEXPECT_FAIL.
We encounter some instability inside V8 / QtScript implementation. We
didn't pass the test before for the same reason. Anyway the test is
broken from conceptual point of view as it touches undefined behavior.
We are not sure what should happen if an JS code is executed in a
context that is about to be rolled up (an exception is thrown).

Reviewed-by: Olivier Goffart