Fixed compile of tst_qscriptextensionplugin on some Windows configurations
The debug and release versions of staticplugin are qmake'd with the
same destination directory.  This causes the generated staticplugin .prl
file to always refer to the debug versions of Qt libraries, even if Qt
was configured with -release.

The .prl mechanism is not useful for this test, so simply disable it to
solve the problem.

This commit is cherry-picked from Qt 4 into Qt 5.  The problem was
hidden until recently in Qt 5 due to a build system bug in qtbase, fixed
by 3225942dbb7faaeff477e137af9c3d78fcaf518e.

(cherry picked from commit f5a63feb8953799de7e787f333575ee37fae8a3f)

Change-Id: I4e88bb0944f83a101445063d59298e49f92451a2
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>