Fix QScriptEngine::ExcludeDeleteLater and metaobject-related autotests
ba635d7e74472f3a54c0c4686966af46d9035c6f in qtbase adds a signal to
QObject.  This breaks some code in QtScript which hardcoded the list of
expected metamethods on QObject, and some other code which hardcoded the
index of the deleteLater slot.  Fix it.

This is a minimal fix which does not address the underlying fragility.

Task-number: QTBUG-23502
Change-Id: Id1269f9d8af01e12587ba8ff2fecf942a7231aff
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>
Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>