Fix JS to QVariant type conversion
1) The special case for LastType (QVariant) should be done before
the general conversion, to avoid calling convertValue() with
targetType == -1 (which is useless and causes a qWarning).

2) moc is about to be changed to use QMetaType::QVariant instead of
QVariant::LastType to represent the QVariant type. But to keep the
CI happy in the transition period we have to check for both. The
LastType check will be removed once qtbase has been updated.

3) Get rid of the variantFromValue function in qscriptqobject. It
was identical to QScriptEnginePrivate::jscValueToVariant.

Change-Id: Ie418facc06c6c7308bc60a3ff66b9a78b109d3d3
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@nokia.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>