Port qtscript to QMetaMethod-based connectNotify()
The const char *-based API is deprecated and will be removed in Qt5.

Also improve the autotest so we check precisely the number of
connect/disconnect notifications received.

The hacks to manually call the notify functions aren't needed
anymore, because the internal QMetaObject::(dis)connect() now does
it for us.

Change-Id: I2e5efed34ee1cba0aad8e60ea00dc9bc2a25bc6c
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>