Revert "JSCore build fix for WEC7"
It's reverted, as there seems to be at least two
implementations of the stl in WEC7. One for each BSP SDK
and one shipped with the compiler, which is in this
case Visual Studio 2008. The Compiler has a stl with std::min
etc. Visual Studio GUI also uses the compiler version of
the stl first. It seems to be that the environment was
setup wrongly. Checksdk also uses the right order.

This reverts commit 8b65d84145f0f415d0963caa7b49da07ac67148d.

Change-Id: Ic00d87777e822d9a7c2ee16ce60cad3782293245
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>