Check that property descriptor members are valid before using them
Even if getPropertyDescriptor() returns true, it's not guaranteed
that PropertyDescriptor::setter() or PropertyDescriptor::value()
returns a valid JSC value.

This code is in an "#ifdef QT_BUILD_SCRIPT_LIB" block, i.e. a
patch we added on top of the original JSC sources.

The lack of checks caused the getter-in-prototype and
indexed-accessors tests from the V8 test suite to assert in
debug mode.

Task-number: QTBUG-17915
Change-Id: I568d83f2f80b28871ea0d934a8e33078ac8aa3ad
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>