QScriptEngine::pushContext(): Don't inherit parent context's scope
This was a regression introduced in Qt 4.6 (JavaScriptCore-based
backend). pushContext() should always create a context with a "clean"
scope (only the Global Object and the context's own activation object
should be in the scope chain).

The scope chain API is internal, but the wrong behavior could still
be observed e.g. through QScriptEngine::evaluate().

Task-number: QTBUG-18188
Change-Id: I138dabc665d7275fb85d3b5e1b473d56096a989e
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>