Suppress JavaScriptCore 'LEAK' messages at application exit
In debug mode, if a QScriptEngine was never destroyed, JavaScriptCore
would output a message like

LEAK: 120 Structure

on stderr at exit.
This is not terribly helpful to users; also, it's not common for Qt
classes to produce such output. Since JavaScriptCore's only usage of
the logging facility is for reporting these "leaks", we can disable
logging altogether.

Task-number: QTBUG-18201
Change-Id: Id598bd1f48129d821a954cc49d8e3ab715d78426
Reviewed-by: Simon Hausmann <simon.hausmann@nokia.com>
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>