Don't disable C++11 on OS X and iOS
On OS X and iOS, enabling C++11 has the side effect of switching
the standard library from libstdc++ to libc++. Qt 5.2 enables
C++11 by default, which leads to a Qt build that uses both
standard libraries. This is something we would like to avoid.
As with clang it is possible to build webkit also with C++11
instead of forcing the use of libc++ also without C++11
we simply build with C++11 on OS X and iOS.

Change-Id: I0ad8fb550ef79ac37ac97a75799ac38aa19d39d4
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>