Fix crash when creating a QScriptEngine in a native thread
This is technicially a cherry-pick of
4e73732d3e72ea59d39ec5a5c01c2e76bbff7dc1 from 4.8. Somehow this change did
not end up in Qt 5. The details below is a copy paste from the original
change.

The change in http://trac.webkit.org/changeset/48412/ introduced a fix
to avoid leaking thread specific data by ensuring get() on ThreadSpecific
works even during the thread destruction phase. The fix worked by setting
the local data again.

However as we can see in the backtrace from QTBUG-22926, the local data
should not be set unconditionally, otherwise our destroy function will be
called recursively when the local data is still set.

Task-number: QTBUG-22926
Change-Id: If59d7e92998b9f1da98c05b94e7b6c725ada4a75
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>