Fix QtScript on 64-bit systems with gcc 8
Similar to commit 92836d052efb6d8073136e8507083f93fb60bb80 in
qtdeclarative, we must use memcpy to do a bit-wise conversion from the
JSCell *m_ptr to an intptr_t for JSValue tag operations.

This fixes the referenced task and all the failing auto-tests.

The most visible result was that

    if ((jsvalue.immediateValue() & 0xffff000000000000ll) == 0xffff000000000000ll)
        return static_cast<int32_t>(jsvalue.immediateValue());

was "optimized" to "return 0", breaking QScriptValue::toNumber() for
integer encoded values.

[ChangeLog][QtScript] Fix incorrect script evaluations on 64-bit systems
with gcc 8.

Change-Id: I525b6b66d0e3c5163ad7e338fd0e866cdf620dad
Task-number: QTBUG-68367
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>