Fix xml output to have valid UTF-8
Some datatags had characters outside the ASCII range. (This is
discouraged).
And when dumping the datatag to xml output we dumped the binary content of
the data tag without encoding it from latin1 to utf-8.  This meant that
all characters in the range 0x80-0xff produced invalid utf-8 sequences.

The solution is to encode the data tag as UTF-8.

Change-Id: Iaaecc992b756836845bfb98429f20f1fb5e2b291
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>
(cherry picked from commit 1d2294735536c53fca8e56ed4b262c4bca6b4d29)
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>