Android: Use -marm instead of -mthumb for armv7 to avoid a crash
Since the Android toolchain has problems with using -mthumb for QtScript
then we switch it to using -marm instead in order to prevent problems
from occurring as a result. Since this is only a problem for the armv7
architecture then we only do this for that architecture.

Change-Id: Ia621f066b8336bb12a2ec1aa99e3a9f9afca186d
Done-with: Markus Maier <markus.maier.sw@gmail.com>
Fixes: QTBUG-67936
Reviewed-by: Markus Maier <markus.maier.sw@gmail.com>
Reviewed-by: BogDan Vatra <bogdan@kdab.com>