Rename VERSION to VERSION.TXT
On case-insensitive file systems, this causes an issue with recent
versions of the Android NDK, because, since Android NDK r19, the libc++ header
"cstddef" contains

  #include <version>

which will resolve to this file instead of the intended header due
to the ordering of the search paths.

The simplest solution is just to give this a suffix so that it can
be differentiated from the headers.

Task-number: QTQAINFRA-2568
Change-Id: Ic1502e65de3dba912ed079d406a65ebf9b935694
Reviewed-by: BogDan Vatra <bogdan@kdab.com>