QScriptValueIterator: replace a QLinkedList with a std::list
The payload type apparently doesn't like to be destroyed unless some
magic foo is enabled (cf. ~QScriptValueIteratorPrivate()). Since the
code also allows to remove items in the middle, we need a container
that guarantees no copies (thus, deletions) happen at all. That leaves
a vector<unique_ptr>, or a std::list. Since this is a deprecated
module, use the least-intrusive option: std::list

Change-Id: Iaadc71959814959bf7ba1a225f375d64edbd9785
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@qt.io>