Fix failing ES test for Georgian script upper-case
Unicode 11 added upper-case letters for Georgian, which didn't exist
when this test was written. Consequently, the test fails now that
we've updated past Unicode 11. The upstream ES test repository has
commented out the test.

Narrowed the test to the block on which it does work, commented on the
two other characters for which it works.

Fixes: QTBUG-79682
Change-Id: I8dd388317cabd62fc5d33487aeb5266d8b9b132d
Reviewed-by: Lars Knoll <lars.knoll@qt.io>